move code to proper file
